Major refactor of compendium_to_card
The CLI has been updated but is still compatible with the prior
versions.
This refactor is the groundwork to support conversion of all entities
from all compendium categories, although only the 'item' category is
currently supported.
A tests dir has been added, but no tests have been written yet!
This version supports importing as a module to allow one to fool around
with the compendium objects.